backend: Add database backup/restore scripts
The scripts expect to operate inside a docker container, the name is
passed as a positional argument to the scripts, and they optionally
support compression (with zstd) and encryption (with gpg).

The restore script is written with argbash to allow for better command
line argument parsing, the main file ends with `.m4` and needs to be
processed after modification before being executed, instructions on how
to do so are present in the file and the generated script is also
included in this commit.